+++ libs/repo-tools/src/generators/core-component/schema.d.ts
@@ -1,0 +1,4 @@
+export interface CoreComponentGeneratorSchema {
+  name: string
+  componentType: 'component' | 'primitive'
+}